fix: give init value valid platform data
In update to bgfx, the bgfx init call attempts to build a valid context.
In a recent update, context will check if the given init is headless by
checking if a valid backplane was provided.  In the workflow for bigg
however, the platform data is set outside of the given init.

By duplicating the platform data, the context is built successfully, and
the pointers are not invalidated