feat: include ImPlot
ImPlot is a nice graphing library on top of ImGui. However,
ImPlot requires creating and destroying a context at the same
time the ImGui context is created and destroyed.

The Bigger framework includes ImPlot, but bigger does not have
the access and control to create and destroy an ImPlot context
at the right place and time.

So this commit adds ImPlot to bigg.